% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip.R
\name{annotatePeaks}
\alias{annotatePeaks}
\title{peak annotate function
this function find nearest gene asscosiated with given peaks, deault database is human.gencodeV19}
\usage{
annotatePeaks(peaks,
  genedb = "~/amber3/no_back_up/data/R_data/human.gencodeV19",
  peak.type = "BED", by_gene = TRUE, min.distance = 2000, pro.len = 500,
  pro = TRUE, subchr = FALSE, head = FALSE)
}
\arguments{
\item{peaks}{peak file, could be bed file path or Grange object, default bed fiel path}
}
\description{
peak annotate function
this function find nearest gene asscosiated with given peaks, deault database is human.gencodeV19
}
\examples{
annotatePeaks()
}

