% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bedtools.coveragebed.range}
\alias{bedtools.coveragebed.range}
\title{calculate overlap % between 2 bed files, need bedtools installed}
\usage{
bedtools.coveragebed.range(functionstring = "coverageBed", bed1, bed2,
  opt.string = "", mcol = TRUE)
}
\arguments{
\item{bed1}{first bed object}

\item{bed2}{second bed object}
}
\description{
calculate overlap % between 2 bed files, need bedtools installed
}
\examples{
bedtools.coveragebed.range()
}

