% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bedtools.coveragebed}
\alias{bedtools.coveragebed}
\title{calculate overlap % between 2 bed files, need bedtools installed}
\usage{
bedtools.coveragebed(functionstring = "coverageBed", bed1, bed2,
  opt.string = "")
}
\arguments{
\item{bed1}{first bed file}

\item{bed2}{second bed file}
}
\description{
calculate overlap % between 2 bed files, need bedtools installed
}
\examples{
bedtools.coveragebed()
}

