% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip.R
\name{cal.rpkm.chip}
\alias{cal.rpkm.chip}
\title{this function calculate RPKM for chip-seq data, support bam and bed input format}
\usage{
cal.rpkm.chip(read, range, input = "bam", pair = F)
}
\arguments{
\item{read}{input file, could be bam or bed format}

\item{range}{GRange object}
}
\description{
this function calculate RPKM for chip-seq data, support bam and bed input format
}
\examples{
cal.rpkm.chip()
}

