% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip.R
\name{diff_peaks_replicate}
\alias{diff_peaks_replicate}
\title{merge peaks based on replicates and find differentially peaks between 2 samples}
\usage{
diff_peaks_replicate(peak1 = NULL, peak2 = NULL, peak3 = NULL,
  peak4 = NULL, min.dis = 10000, min.dis.for.diff = 10000,
  remove.small0 = 0, remove.small0.for.diff = 0,
  remove.small.for.diff = NULL, remove.small = NULL,
  remove.small.each.replicate = NULL, remove.small2 = 0, name1 = NULL,
  name2 = NULL, name3 = NULL)
}
\arguments{
\item{peak1}{sample1 rep1}

\item{peak2}{sample1 rep2}

\item{peak3}{sample2 rep1}

\item{peak4}{sample2 rep2}
}
\description{
merge peaks based on replicates and find differentially peaks between 2 samples
}
\examples{
diff_peaks_replicate()
}

