% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{generate.sliding.window}
\alias{generate.sliding.window}
\title{generate sliding window ranges}
\usage{
generate.sliding.window(region, step.percent = 0.01, step.base = 10,
  up = 1000, down = 1000)
}
\arguments{
\item{region}{GRange object}
}
\description{
generate sliding window ranges
}
\examples{
generate.sliding.window()
}

