% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip.R
\name{get_common_peak3}
\alias{get_common_peak3}
\title{extract common peaks based on replicates}
\usage{
get_common_peak3(peak1, peak2, return.range = FALSE, min.overlap = 0.5)
}
\arguments{
\item{min.overlap}{overlap % cutoff between two files, default 0.5}

\item{read1}{rep1}

\item{read1}{rep2}
}
\description{
extract common peaks based on replicates
}
\examples{
get_common_peak3()
}

