% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip.R
\name{merge_peaks}
\alias{merge_peaks}
\title{merge neighbouring peaks to large one by given distance}
\usage{
merge_peaks(peak = peakfile, min.dis = 10000, remove.small = NULL,
  remove.small0 = 0, return.range = FALSE)
}
\arguments{
\item{peak}{peak bed file path}

\item{mis.dis}{distance cutoff, default 10000}
}
\description{
merge neighbouring peaks to large one by given distance
}
\examples{
merge_peaks()
}

