% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth.R
\name{meth_count_region_for_bsseq}
\alias{meth_count_region_for_bsseq}
\title{calculate methylation level based on given Grange from bsseq object}
\usage{
meth_count_region_for_bsseq(BS, range, min.cov = 5, min.cpg.num = 1)
}
\description{
calculate methylation level based on given Grange from bsseq object
}
\examples{
meth_count_region_for_bsseq()
}

