% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth.R
\name{plot_meth_sliding_bsseq}
\alias{plot_meth_sliding_bsseq}
\title{plot and calculate methylation along sliding windows from bsseq object
this function plot and calculate slidng windows methylation}
\usage{
plot_meth_sliding_bsseq(BS.BM, region, win.base = 5000, step.base = 5000,
  up.base = 2e+05, down.base = 2e+05, win.percent = 0.01,
  step.percent = 0.01, up.percent = 1, down.percent = 1, min.num = 20,
  min.mc.depth = 0, frame = TRUE, main = "DNAm", col = "black",
  append = FALSE, dist = "base", x.axis = TRUE, ylim = c(0, 1))
}
\arguments{
\item{BS.BM}{bsseq object}

\item{region}{GRange object}
}
\description{
plot and calculate methylation along sliding windows from bsseq object
this function plot and calculate slidng windows methylation
}
\examples{
plot_meth_sliding_bsseq()
}

